eventdev: fix memory size for telemetry
[ upstream commit e0fb0051ff8f621a2b001d9243cebe50f3d42cf5 ]

Fix the invalid memory free operation caused by extending
xstats ID width from 32 to 64 bits.

Telemetry application crashed when freeing memory,
hence, the telemetry app could not display xstats.

Fixes: 1bdfe4d76e98 ("eventdev: increase xstats ID width to 64 bits")

Acked-by: Jerin Jacob <jerinj@marvell.com>

+++ lib/eventdev/rte_eventdev.c
@@ -1678,7 +1678,7 @@ eventdev_build_telemetry_data(int dev_id,
 	if (xstat_names == NULL)
 		return -1;
 
-	ids = malloc((sizeof(unsigned int)) * num_xstats);
+	ids = malloc((sizeof(uint64_t)) * num_xstats);
 	if (ids == NULL) {
 		free(xstat_names);
 		return -1;
